requirements.txt
easydict
ipython
numpy
opencv-python
Pillow
PyYAML
scikit-learn
torch+cu111
torchaudio
torchvision+cu111
PyYAML
easydict
scikit-learn    
tqdm